Native position changed from float to double

+++ src/main/java/uk/co/caprica/vlcj/media/ThumbnailApi.java
@@ -72,7 +72,7 @@ public final class ThumbnailApi extends BaseApi {
      * @param timeout timeout for thumbnail generation
      * @return thumbnail request
      */
-    public ThumbnailRequest requestByPosition(float position, ThumbnailerSeekSpeed speed, int width, int height, boolean crop, PictureType pictureType, long timeout) {
+    public ThumbnailRequest requestByPosition(double position, ThumbnailerSeekSpeed speed, int width, int height, boolean crop, PictureType pictureType, long timeout) {
         libvlc_media_thumbnail_request_t request = libvlc_media_thumbnail_request_by_pos(libvlcInstance, mediaInstance, position, speed.intValue(), width, height, crop ? 1 : 0, pictureType.intValue(), timeout);
         if (request != null) {
             return new ThumbnailRequest(request);
